cloud-brain/requirements-dev.txt
# Testing frameworks
pytest==9.0.2
pytest-cov==7.0.0
pytest-asyncio==1.3.0
hypothesis==6.151.6

# Mocking
moto==5.1.21

# Code quality
ruff==0.15.1
mypy==1.19.1
boto3-stubs[essential]==1.42.48
